fix: 修复直推 commit 82b841b7 的 CI 红灯——测试对齐 + useAgent 拆分至限额内
- userMessageBubbleSkillFlowSource：修正误删残留的双星号正则（语法错误）；
  serif 特性已给 chip 名称节点加 font-serif，过时的 not 断言改为 pin 现状
- modelSelectorTriggerColor：补 pin 模型名 font-serif（82b841b7 漏更新）
- historyScrollPerformanceSource：isLastMessage 正则兼容 prettier 换行；
  renderItems 定义方式变化（useMemo→直取 messages）同步更新
- chatViewHistoryPaginationSource：分页断言改读新 hook 文件
- useAgent.ts 1130→988 行（>1000 触发 Large File Check）：按特性内聚拆出
  useHistoryTracePagination（分页编排）/ extractSessionConfig /
  useAutoModeSetting，行为逐字搬移不变；补齐 exhaustive-deps 依赖
