+++ frontend/src/hooks/useAgent.ts
@@ -6,12 +6,7 @@
 import { useState, useCallback, useRef, useEffect } from "react";
 import toast from "react-hot-toast";
 import i18n from "../i18n";
-import type {
-  Message,
-  ConnectionStatus,
-  MessageAttachment,
-  Feedback,
-} from "../types";
+import type { Message, ConnectionStatus, MessageAttachment } from "../types";
 import { sessionApi, type BackendSession } from "../services/api";
 import { feedbackApi } from "../services/api/feedback";
 import { useAuth } from "../hooks/useAuth";
@@ -60,12 +55,10 @@ import {
   applyFeedbackToMessages,
   resolveHistoryStreamRunId,
 } from "./useAgent/historyLoadState";
-import {
-  HISTORY_TRACE_PAGE_SIZE,
-  canLoadOlderHistory,
-  mergeOlderHistoryEvents,
-  resolveHistoryTraceWindow,
-} from "./useAgent/historyPagination";
+import { HISTORY_TRACE_PAGE_SIZE } from "./useAgent/historyPagination";
+import { useHistoryTracePagination } from "./useAgent/historyTracePagination";
+import { extractSessionConfig } from "./useAgent/sessionConfig";
+import { useAutoModeSetting } from "./useAgent/autoModeSetting";
 
 export function useAgent(options?: UseAgentOptions): UseAgentReturn {
   const { hasAnyPermission } = useAuth();
@@ -80,8 +73,6 @@ export function useAgent(options?: UseAgentOptions): UseAgentReturn {
   const [isLoading, setIsLoading] = useState(false);
   const [isLoadingHistory, setIsLoadingHistory] = useState(false);
   const [historyLoadGeneration, setHistoryLoadGeneration] = useState(0);
-  const [hasMoreHistoryTraces, setHasMoreHistoryTraces] = useState(false);
-  const [isLoadingOlderHistory, setIsLoadingOlderHistory] = useState(false);
   const [sessionId, setSessionId] = useState<string | null>(null);
   const [currentProjectId, setCurrentProjectId] = useState<string | null>(null);
   const [error, setError] = useState<string | null>(null);
@@ -98,22 +89,7 @@ export function useAgent(options?: UseAgentOptions): UseAgentReturn {
     Record<string, ActiveGoalSpec>
   >({});
   const [goalModeEnabled, setGoalModeEnabled] = useState(false);
-  const [autoModeEnabled, setAutoModeEnabled] = useState(() => {
-    try {
-      return localStorage.getItem("lamb-chat-auto-mode") === "true";
-    } catch {
-      return false;
-    }
-  });
-
-  // Persist autoModeEnabled to localStorage
-  useEffect(() => {
-    try {
-      localStorage.setItem("lamb-chat-auto-mode", String(autoModeEnabled));
-    } catch {
-      /* storage unavailable */
-    }
-  }, [autoModeEnabled]);
+  const [autoModeEnabled, setAutoModeEnabled] = useAutoModeSetting();
 
   // Refs for connection management
   const abortControllerRef = useRef<AbortController | null>(null);
@@ -135,16 +111,6 @@ export function useAgent(options?: UseAgentOptions): UseAgentReturn {
 
   // Track last event timestamp from history
   const lastHistoryTimestampRef = useRef<Date | null>(null);
-
-  // History trace-window pagination state (older pages prepend on scroll)
-  const historyEventsRef = useRef<HistoryEvent[]>([]);
-  const historyTraceWindowRef = useRef<{
-    oldest_trace_started_at: string;
-    oldest_trace_id: string;
-  } | null>(null);
-  const historyFeedbackRef = useRef<Feedback[]>([]);
-  const hasMoreHistoryTracesRef = useRef(false);
-  const isLoadingOlderHistoryRef = useRef(false);
 
   // Subagent tracking stack
   const activeSubagentStackRef = useRef<SubagentStackItem[]>([]);
@@ -206,6 +172,25 @@ export function useAgent(options?: UseAgentOptions): UseAgentReturn {
   useEffect(() => {
     messagesRef.current = messages;
   }, [messages]);
+
+  // History trace-window pagination (older pages prepend on scroll)
+  const {
+    hasMoreHistoryTraces,
+    isLoadingOlderHistory,
+    recordFirstWindow,
+    recordFeedback,
+    loadOlderHistory,
+    reset: resetHistoryPagination,
+  } = useHistoryTracePagination({
+    options,
+    sessionIdRef,
+    isLoadingHistoryRef,
+    processedEventIdsRef,
+    messagesRef,
+    streamingMessageIdRef,
+    setMessages,
+    setGoalsByRunId,
+  });
 
   // Create event handler context
   const createEventHandlerContext = useCallback(
@@ -315,11 +300,7 @@ export function useAgent(options?: UseAgentOptions): UseAgentReturn {
 
       processedEventIdsRef.current.clear();
       lastHistoryTimestampRef.current = null;
-      historyEventsRef.current = [];
-      historyTraceWindowRef.current = null;
-      historyFeedbackRef.current = [];
-      hasMoreHistoryTracesRef.current = false;
-      setHasMoreHistoryTraces(false);
+      resetHistoryPagination();
       void sessionApi.markRead(targetSessionId).catch(() => {});
       const feedbackPromise = canReadFeedback
         ? feedbackApi
@@ -374,42 +355,11 @@ export function useAgent(options?: UseAgentOptions): UseAgentReturn {
             (sessionData.metadata?.current_run_id as string) ||
             null;
 
-          // 从 metadata 提取配置信息
-          const sessionConfig = {
-            agent_id: (sessionData.metadata?.agent_id as string) || undefined,
-            agent_options:
-              (sessionData.metadata?.agent_options as Record<
-                string,
-                boolean | string | number
-              >) || undefined,
-            disabled_tools:
-              (sessionData.metadata?.disabled_tools as string[]) || undefined,
-            disabled_skills:
-              (sessionData.metadata?.disabled_skills as string[]) || undefined,
-            enabled_skills:
-              (sessionData.metadata?.enabled_skills as string[]) || undefined,
-            persona_preset_id:
-              (sessionData.metadata?.persona_preset_id as string) || undefined,
-            persona_preset_name:
-              (sessionData.metadata?.persona_preset_name as string) ||
-              undefined,
-            persona_snapshot:
-              (sessionData.metadata?.persona_snapshot as
-                | import("../types").PersonaPresetSnapshot
-                | undefined) || undefined,
-            disabled_mcp_tools:
-              (sessionData.metadata?.disabled_mcp_tools as string[]) ||
-              undefined,
-            team_id: (sessionData.metadata?.team_id as string) || undefined,
-          };
+          const sessionConfig = extractSessionConfig(sessionData);
           setGoalModeEnabled(false);
 
           const historyEvents = (eventsData.events || []) as HistoryEvent[];
-          historyEventsRef.current = historyEvents;
-          const firstWindowState = resolveHistoryTraceWindow(eventsData);
-          historyTraceWindowRef.current = firstWindowState.traceWindow;
-          hasMoreHistoryTracesRef.current = firstWindowState.hasMore;
-          setHasMoreHistoryTraces(firstWindowState.hasMore);
+          recordFirstWindow(eventsData);
           let reconstructedMessages = reconstructMessagesFromEvents(
             historyEvents,
             processedEventIdsRef.current,
@@ -444,7 +394,7 @@ export function useAgent(options?: UseAgentOptions): UseAgentReturn {
 
           void feedbackPromise.then((feedbackList) => {
             if (!feedbackList || isStaleHistoryLoad()) return;
-            historyFeedbackRef.current = feedbackList.items;
+            recordFeedback(feedbackList.items);
             setMessages((previous) =>
               applyFeedbackToMessages(previous, feedbackList.items),
             );
@@ -499,97 +449,11 @@ export function useAgent(options?: UseAgentOptions): UseAgentReturn {
       canReadFeedback,
       clearSteerMessages,
       hydrateSteers,
+      recordFirstWindow,
+      recordFeedback,
+      resetHistoryPagination,
     ],
   );
-
-  // Load one more page of older runs and prepend the rebuilt messages.
-  // Runs mid-stream are tolerated: the streaming run's assistant message is
-  // re-marked streaming after the rebuild so SSE updates keep applying.
-  const loadOlderHistory = useCallback(async () => {
-    const traceWindow = historyTraceWindowRef.current;
-    if (!traceWindow || !sessionIdRef.current) return;
-    const targetSessionId = sessionIdRef.current;
-    if (
-      !canLoadOlderHistory({
-        sessionId: targetSessionId,
-        traceWindow,
-        hasMore: hasMoreHistoryTracesRef.current,
-        isLoading:
-          isLoadingOlderHistoryRef.current || isLoadingHistoryRef.current,
-      })
-    ) {
-      return;
-    }
-    isLoadingOlderHistoryRef.current = true;
-    setIsLoadingOlderHistory(true);
-    try {
-      const olderData = await sessionApi.getEvents(targetSessionId, {
-        // 与首屏一致：带上活动 run 语义，避免深页因 status 过滤
-        // 丢掉写入方已死（stale running）的历史轮次
-        include_active_user_message: true,
-        compact_message_chunks: true,
-        trace_limit: HISTORY_TRACE_PAGE_SIZE,
-        before_trace_started_at: traceWindow.oldest_trace_started_at,
-        before_trace_id: traceWindow.oldest_trace_id,
-      });
-      if (sessionIdRef.current !== targetSessionId) return;
-
-      const olderEvents = (olderData.events || []) as HistoryEvent[];
-      if (olderEvents.length === 0) {
-        hasMoreHistoryTracesRef.current = false;
-        setHasMoreHistoryTraces(false);
-        return;
-      }
-
-      const mergedEvents = mergeOlderHistoryEvents(
-        olderEvents,
-        historyEventsRef.current,
-      );
-      historyEventsRef.current = mergedEvents;
-      const windowState = resolveHistoryTraceWindow(olderData);
-      historyTraceWindowRef.current = windowState.traceWindow;
-      hasMoreHistoryTracesRef.current = windowState.hasMore;
-      setHasMoreHistoryTraces(windowState.hasMore);
-
-      // 全量重建（而非仅重建旧页）：同一 run 可能跨多条 trace（重试回放），
-      // 按页独立重建会产生重复消息 id。重建使用全新的 subagent 栈。
-      let reconstructedMessages = reconstructMessagesFromEvents(
-        mergedEvents,
-        processedEventIdsRef.current,
-        { options, activeSubagentStack: [] },
-      );
-      const streamingRunId =
-        messagesRef.current.find(
-          (message) =>
-            message.role === "assistant" &&
-            message.isStreaming &&
-            message.runId,
-        )?.runId ?? null;
-      if (streamingRunId) {
-        const prepared = prepareMessagesForRunningRun(
-          reconstructedMessages,
-          streamingRunId,
-          undefined,
-          messagesRef.current,
-        );
-        reconstructedMessages = prepared.messages;
-        streamingMessageIdRef.current = prepared.streamingMessageId;
-      }
-      if (historyFeedbackRef.current.length > 0) {
-        reconstructedMessages = applyFeedbackToMessages(
-          reconstructedMessages,
-          historyFeedbackRef.current,
-        );
-      }
-      setGoalsByRunId(extractGoalsByRunFromEvents(mergedEvents));
-      setMessages(reconstructedMessages);
-    } catch (err) {
-      console.warn("[loadOlderHistory] Failed to load older history:", err);
-    } finally {
-      isLoadingOlderHistoryRef.current = false;
-      setIsLoadingOlderHistory(false);
-    }
-  }, [options]);
 
   // Send message
   const sendMessage = useCallback(
@@ -998,13 +862,7 @@ export function useAgent(options?: UseAgentOptions): UseAgentReturn {
     setConnectionStatus("disconnected");
     processedEventIdsRef.current.clear();
     lastHistoryTimestampRef.current = null;
-    historyEventsRef.current = [];
-    historyTraceWindowRef.current = null;
-    historyFeedbackRef.current = [];
-    hasMoreHistoryTracesRef.current = false;
-    setHasMoreHistoryTraces(false);
-    isLoadingOlderHistoryRef.current = false;
-    setIsLoadingOlderHistory(false);
+    resetHistoryPagination();
     streamingMessageIdRef.current = null;
     sessionIdRef.current = null;
     currentRunIdRef.current = null;
@@ -1017,7 +875,7 @@ export function useAgent(options?: UseAgentOptions): UseAgentReturn {
       abortControllerRef.current = null;
     }
     clearReconnectTimeout(reconnectTimeoutRef);
-  }, [clearSteerMessages]);
+  }, [clearSteerMessages, resetHistoryPagination]);
 
   const clearActiveGoal = useCallback(() => {
     setGoalModeEnabled(false);
